// frontend/src/api/model/methodologicalStageResponseDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { StageDeliverableResponseDto } from './stageDeliverableResponseDto.ts';

export interface MethodologicalStageResponseDto {
  id?: string;
  processId?: string;
  order?: number;
  code?: string;
  displayName?: string;
  status?: string;
  startedAt?: string;
  closedAt?: string;
  deliverables?: StageDeliverableResponseDto[];
}